+++ pricing/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from pricing.factory import get_provider
+
+__all__ = ['get_provider']